refactor(dart): extract TypeIds into standalone meta/type_ids.dart
Move TypeIds out of type_resolver.dart into its own file under
lib/src/meta/ to eliminate unnecessary dependency on the heavy
type_resolver module from low-level files like field_type.dart,
hash_util.dart, primitive_serializers.dart, and
typed_array_serializers.dart.
