+++ dart/packages/fory/lib/src/serializer/primitive_serializers.dart
@@ -1,6 +1,6 @@
 import 'package:fory/src/context/read_context.dart';
 import 'package:fory/src/context/write_context.dart';
-import 'package:fory/src/resolver/type_resolver.dart';
+import 'package:fory/src/meta/type_ids.dart';
 import 'package:fory/src/serializer/serializer.dart';
 import 'package:fory/src/types/fixed_ints.dart';
 import 'package:fory/src/types/float16.dart';
