+++ dart/packages/fory/lib/src/serializer/typed_array_serializers.dart
@@ -2,7 +2,7 @@ import 'dart:typed_data';
 
 import 'package:fory/src/context/read_context.dart';
 import 'package:fory/src/context/write_context.dart';
-import 'package:fory/src/resolver/type_resolver.dart';
+import 'package:fory/src/meta/type_ids.dart';
 import 'package:fory/src/serializer/serializer.dart';
 
 void writeTypedArrayBytes(
